requirements_garage-habitat.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.1.0=pypi_0
aiosignal=1.2.0=pypi_0
akro=0.0.8=pypi_0
attrs=21.4.0=pyhd8ed1ab_0
blas=1.0=mkl
braceexpand=0.1.7=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.6.15=ha878542_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py37h89c1867_0
cffi=1.15.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
cloudpickle=2.0.0=pypi_0
cma=2.7.0=pypi_0
colorama=0.4.5=pyhd8ed1ab_0
cpuonly=1.0=0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.28=pypi_0
decorator=4.4.2=pypi_0
distlib=0.3.4=pypi_0
dm-control=1.0.2=pypi_0
dm-env=1.5=pypi_0
dm-tree=0.1.7=pypi_0
dowel=0.0.3=pypi_0
fasteners=0.17.3=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.7.1=pypi_0
fonttools=4.33.3=pypi_0
freetype=2.10.4=h0708190_1
frozenlist=1.3.0=pypi_0
future=0.18.2=pypi_0
garage=2020.9.0rc2.dev0=dev_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
glfw=2.5.3=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.9.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.43.0=pypi_0
gym=0.23.0=pypi_0
gym-miniworld=2020.1.9=dev_0
gym-notices=0.0.7=pypi_0
habitat=0.2.2=dev_0
habitat-sim=0.2.2=pypi_0
habitat-sim-mutex=1.0=display_bullet
icu=67.1=he1b5a44_0
idna=3.3=pypi_0
ifcfg=0.22=pypi_0
imageio=2.19.3=pypi_0
imageio-ffmpeg=0.4.7=pyhd8ed1ab_0
importlib-metadata=4.11.4=pypi_0
importlib-resources=5.8.0=pypi_0
iniconfig=1.1.1=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
jpeg=9b=h024ee3a_2
jsonschema=4.6.0=pypi_0
kiwisolver=1.4.2=pypi_0
labmaze=1.0.5=pypi_0
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.8.0=21_mkl
libcblas=3.8.0=21_mkl
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h516909a_0
liblapack=3.8.0=21_mkl
libllvm10=10.0.1=he513fc3_3
libpng=1.6.37=h21135ba_2
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.0.9=he6b73bb_1
libuv=1.42.0=h7f98852_0
libxcb=1.13=h7f98852_1003
llvmlite=0.38.1=pypi_0
lmdb=1.3.0=pypi_0
lxml=4.8.0=pypi_0
markdown=3.3.7=pypi_0
matplotlib=3.5.2=pypi_0
matplotlib-base=3.2.2=py37h1d35a4c_1
mkl=2020.2=256
mkl-service=2.3.0=py37h8f50634_2
mkl_fft=1.2.0=py37h161383b_1
mkl_random=1.2.0=py37h9fdb41a_1
moviepy=2.0.0.dev2=pypi_0
msgpack=1.0.4=pypi_0
mujoco=2.1.5=pypi_0
mujoco-py=2.0.2.8=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.6=he412f7d_0
networkx=2.6.3=pypi_0
ninja=1.10.2=h4bd325d_0
numba=0.55.2=pypi_0
numpy=1.18.5=pypi_0
numpy-quaternion=2022.4.2=pypi_0
oauthlib=3.2.0=pypi_0
objectio=0.2.29=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1o=h7f8727e_0
packaging=21.3=pypi_0
pickle5=0.0.12=pypi_0
pillow=9.1.0=pypi_0
pip=21.2.2=py37h06a4308_0
platformdirs=2.5.2=pypi_0
pluggy=1.0.0=pypi_0
proglog=0.1.10=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.1=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
py=1.11.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pyglet=1.5.23=pypi_0
pyopengl=3.1.6=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.1=pypi_0
pytest=7.1.2=pypi_0
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytorch=1.8.1=py3.7_cpu_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
quaternion=2021.7.13.7.29.27=py37h5e8e339_0
ray=1.13.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-image=0.19.3=pypi_0
scipy=1.7.3=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=61.2.0=py37h06a4308_0
simplejson=3.17.6=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=3.0.5=pyh44b312d_0
sqlite=3.38.3=hc218d9a_0
tabulate=0.8.10=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5.1=pypi_0
tifffile=2021.11.2=pypi_0
tk=8.6.11=h1ccaba5_0
tomli=2.0.1=pypi_0
torchaudio=0.8.1=py37
torchvision=0.9.1=py37_cpu
tornado=6.1=py37h5e8e339_1
tqdm=4.64.0=pyhd8ed1ab_0
typer=0.4.1=pypi_0
typing_extensions=4.2.0=pyha770c72_1
urllib3=1.26.9=pypi_0
virtualenv=20.14.1=pypi_0
webdataset=0.1.40=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pypi_0
withbullet=2.0=0
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxcursor=1.2.0=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-libxinerama=1.1.4=h9c3ff4c_1001
xorg-libxrandr=1.5.2=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-randrproto=1.5.0=h7f98852_1001
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7f8727e_1
yacs=0.1.8=pypi_0
zipp=3.8.0=pypi_0
zlib=1.2.12=h7f8727e_2
